pyproject.toml
[project]
name = "mavis-reporting-prototype"
version = "0.1.0"
description = "Prototype code for the commissioner reporting component of Mavis"
license = "MIT"
readme = "README.md"
requires-python = ">=3.13"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
flask = "==3.1.1"
python-dotenv = "==1.1.1"
nhsuk-frontend-jinja = "==0.3.1"
gunicorn = "^23.0.0"
py-healthcheck = "^1.10.1"
boto3 = "==1.39.13"
pyyaml = "^6.0.2"
requests = "^2.32.4"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
ruff = "==0.12.5"
honcho = "^2.0.0"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
